Changing KDNode variable types

+++ reworking/src/indexes/kd_tree/kd_node.cpp
@@ -20,11 +20,11 @@ public:
     shared_ptr<KDNode> left_child;
     shared_ptr<KDNode> right_child;
 
-    int64_t left_position; // Strict smaller than key from this position backwards
-    int64_t right_position; // Greater or equal than key from this position forwards
+    size_t left_position; // Strict smaller than key from this position backwards
+    size_t right_position; // Greater or equal than key from this position forwards
 
 
-    KDNode(size_t column, float key, int64_t left_position, int64_t right_position)
+    KDNode(size_t column, float key, size_t left_position, size_t right_position)
     : key(key), column(column), left_position(left_position), right_position(right_position){
         left_child = nullptr;
         right_child = nullptr;
